+++ plugin/gatsby-plugin-image/index.js
@@ -10,6 +10,8 @@ exports.createGatsbyImageDataResolver = (gatsbyUtils, pluginOptions) => {
 
   if (gatsbyImageResolver) {
     const resolvers = {};
+    // Make the resolver nullable, createGatsbyPluginImageResolver sets the type to 'GatsbyImageData!'
+    gatsbyImageResolver.type = 'GatsbyImageData';
 
     transformTypes.forEach((type) => {
       // Add gatsbyImageData resolver